requirements.txt
invoke
boto3
moto
pytest
pyyaml
